+++ include/sw/dsp/sdr/sdr.hpp
@@ -10,4 +10,5 @@
 // SPDX-License-Identifier: MIT
 
 #include <sw/dsp/sdr/constellation.hpp>
+#include <sw/dsp/sdr/metrics.hpp>
 #include <sw/dsp/sdr/rrc.hpp>